+++ scripts/helpers/tmux_echo_functions.sh
@@ -1,3 +1,7 @@
+_has_emacs_mode_keys() {
+	$(tmux show -gw mode-keys | grep -q emacs)
+}
+
 tmux_echo() {
 	local message="$1"
 	tmux run-shell "echo '$message'"
@@ -12,8 +16,13 @@ echo_err() {
 }
 
 end_message() {
+	if _has_emacs_mode_keys; then
+		local continue_key="ESCAPE"
+	else
+		local continue_key="ENTER"
+	fi
 	tmux_echo ""
 	tmux_echo "TMUX environment reloaded."
 	tmux_echo ""
-	tmux_echo "Done, press ENTER to continue."
+	tmux_echo "Done, press $continue_key to continue."
 }